requirements.txt
opencv-python
numpy
fastapi
uvicorn
python-multipart
cloudinary
python-dotenv
requests
